Add Linux loader
This privdes just enough firmware to load Linux. The loader is linked to
run from the aliased BRAM address. If that changes the linker script and
'make load' address should be updated.

The printf comes from https://github.com/mpredfearn/simple-printf

 1. Build Linux for microwatt
  objcopy -O binary microwatt/vmlinux vmlinux.bin

 2. Build a dtb

 3. Set MW_DEBUG to point to a copy of mw_debug and type `make load` to
    load the loader into BRAM.

 4. Press 'RESET'

 5. Follow the instructions:

	Microwatt Loader (May 12 2020 18:04:22)

	Load binaries into SDRAM and press 'l' to start:

	 mw_debug -b jtag stop load vmlinux.bin load microwatt.dtb 1000000 start

If you want to test a smaller program, load it to the start of SDRAM and
the loader will jump to it.

Signed-off-by: Joel Stanley <joel@jms.id.au>